feat: findallができるようになった

+++ src/domains/mod.rs
@@ -4,19 +4,19 @@ pub mod documents;
 
 #[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, Ord, PartialOrd)]
 pub struct Id<T> {
-    id: u64,
+    id: i32,
     _phantom: PhantomData<T>,
 }
 
 impl<T> Id<T> {
-    pub fn new(id: u64) -> Self {
+    pub fn new(id: i32) -> Self {
         Self {
             id,
             _phantom: PhantomData,
         }
     }
 
-    pub fn get(&self) -> u64 {
+    pub fn get(&self) -> i32 {
         self.id
     }
 }
